feat: implementation of pokemon details - added the connector, page and vm for the pokemon details page - created a DefaultTabController for the views of About, Stats, Moves and Abilities - created another function for pokemon actions for disposing the dirty pokemon details - created constant variable in constant for magic numbers - added a gesture detector to route the user to the details page when Pokemon Card is Tap(onTap) - modified the object type of model in pokemon_details and stats - flutter analyzed no issues found - dart format completed no issues found

+++ lib/api/model/pokemon_details.dart
@@ -10,8 +10,8 @@ part 'pokemon_details.g.dart';
 @freezed
 class PokemonDetails with _$PokemonDetails {
   const factory PokemonDetails({
-    @JsonKey(name: 'height') required double height,
-    @JsonKey(name: 'weight') required double weight,
+    @JsonKey(name: 'height') required int height,
+    @JsonKey(name: 'weight') required int weight,
     @JsonKey(name: 'abilities') required List<Ability> abilities,
     @JsonKey(name: 'moves') required List<Move> moves,
     @JsonKey(name: 'stats') required List<Stats> stats,
